Codechange: Move colour brightness methods to palette code.
This allows reuse outside of blitters.
